Add files via upload
Using a dataset from salaries.ai-jobs.net as of 9.26.2022, we wanted to understand , as future managers, as well as current students looking for jobs, what it may cost to hire data professionals of specific titles and skill levels in various locations. This has potential cost-saving benefits for companies as we will explore in the next section.